requirements.txt
numpy
matplotlib
pillow
tqdm
pycocotools
pandas
boto3
streamlit
google-genai
torch
torchvision
